attempt to fix

+++ include/headerer.h
@@ -16,7 +16,7 @@ typedef struct {
     char magic[4];
     char version[64];
     uint16_t flags;
-    uint32_t code_size;
+    uint64_t code_size;
     uint64_t globalAmount;
 } LeyoHeader;
